Reorganization in tests project
Moved all test classes in root folder to Misc.
