+++ test/Serilog.Sinks.MSSqlServer.Tests/Misc/MiscFeaturesTests.cs
@@ -6,7 +6,7 @@ using Serilog.Sinks.MSSqlServer.Tests.TestUtils;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Serilog.Sinks.MSSqlServer.Tests
+namespace Serilog.Sinks.MSSqlServer.Tests.Misc
 {
     [Trait(TestCategory.TraitName, TestCategory.Integration)]
     public class MiscFeaturesTests : DatabaseTestsBase
@@ -23,7 +23,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
             // meaning of the flag. Contrast with LogEventExcludeStandardProperties below.
 
             // Arrange
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions()
+            var columnOptions = new MSSqlServer.ColumnOptions()
             {
                 AdditionalColumns = new List<SqlColumn>
                 {
@@ -69,7 +69,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
         public void LogEventExcludeStandardColumns()
         {
             // Arrange
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
             columnOptions.Store.Remove(StandardColumn.Properties);
             columnOptions.Store.Add(StandardColumn.LogEvent);
             columnOptions.LogEvent.ExcludeStandardColumns = true;
@@ -107,7 +107,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
         public void ExcludeIdColumn()
         {
             // Arrange
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
             columnOptions.Store.Remove(StandardColumn.Id);
 
             Log.Logger = new LoggerConfiguration()
@@ -136,7 +136,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
         public void BigIntIdColumn()
         {
             // Arrange
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
             columnOptions.Id.DataType = SqlDbType.BigInt;
 
             Log.Logger = new LoggerConfiguration()
@@ -165,7 +165,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
         public void XmlPropertyColumn()
         {
             // Arrange
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
             columnOptions.Properties.DataType = SqlDbType.Xml;
 
             Log.Logger = new LoggerConfiguration()
@@ -202,7 +202,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
             // Arrange
             var schemaName = "CustomTestSchema";
             var tableName = "CustomSchemaLogTable";
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
 
             Log.Logger = new LoggerConfiguration()
                 .WriteTo.MSSqlServer
@@ -235,7 +235,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
             // Arrange
             var schemaName = "CustomTestSchema";
             var tableName = "CustomSchemaLogTable";
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
 
             Log.Logger = new LoggerConfiguration()
                 .WriteTo.MSSqlServer
@@ -264,7 +264,7 @@ namespace Serilog.Sinks.MSSqlServer.Tests
         public void LogEventStoreAsEnum()
         {
             // Arrange
-            var columnOptions = new Serilog.Sinks.MSSqlServer.ColumnOptions();
+            var columnOptions = new MSSqlServer.ColumnOptions();
             columnOptions.Level.StoreAsEnum = true;
             columnOptions.Store.Add(StandardColumn.LogEvent);
